bookmark API 구현완료

+++ back/src/models/user.ts
@@ -2,6 +2,7 @@ import { Schema, Document, Model, ObjectId, model } from 'mongoose';
 
 import { IUserBookmark, UserBookmarkModel } from '@/src/models/userBookmark';
 import Video, { IVideoDocument } from '@/src/models/video';
+import video from '@/src/models/video';
 
 interface IUser {
   snsId: string;
@@ -17,7 +18,10 @@ interface IUser {
 }
 export interface IUserDocument extends IUser, Document {
   createBookmark(title: string): IUserDocument;
-  addBookmark(videoInfo: IVideoDocument, selectBookmarkId: string): string;
+  addBookmark(
+    videoInfo: IVideoDocument,
+    selectBookmarkId: string,
+  ): { success: string[]; fail: string[] };
   removeBookmark(bookmarkId: string): void;
   removeYoutube(bookmarkId: string, videoId: string): void;
 }
@@ -87,14 +91,13 @@ userSchema.methods.addBookmark = async function (
 ) {
   console.log('methods.addBookmark');
 
-  await selectBookmarkId.forEach(async (id: string) => {
-    const foundBookmark = await UserBookmarkModel.findById(id);
-
+  await selectBookmarkId.forEach(async (bmId: string) => {
+    const foundBookmark = await UserBookmarkModel.findById(bmId);
     const existedVideo = foundBookmark?.videos.findIndex(
       (v) => v.videoId.toString() === videoInfo._id.toString(),
     );
 
-    if (existedVideo! < 0) {
+    if (0 > existedVideo!) {
       foundBookmark!.videos.push({ videoId: videoInfo._id });
       foundBookmark!.count += 1;
       await foundBookmark?.save();
@@ -103,12 +106,20 @@ userSchema.methods.addBookmark = async function (
 };
 
 userSchema.methods.removeBookmark = async function (bookmarkId: string) {
+  const foundBookmark = await UserBookmarkModel.findById(bookmarkId);
+  foundBookmark?.videos.forEach(async ({ videoId }) => {
+    const result = await UserBookmarkModel.findOne(videoId);
+    if (!result) await Video.findByIdAndRemove(videoId);
+  });
+
+  // 유저 북마크 삭제
   const updatedBookmark = this.bookmarks.bookmark.filter(
     (bm: ObjectId) => bm.toString() !== bookmarkId.toString(),
   );
   this.bookmarks = { bookmark: updatedBookmark };
   this.save();
 
+  // 북마크 도큐먼트 삭제
   await UserBookmarkModel.findByIdAndRemove(bookmarkId);
 };
 
@@ -117,15 +128,21 @@ userSchema.methods.removeYoutube = async function (
   videoId: string,
 ) {
   const foundBookmark = await UserBookmarkModel.findById(bookmarkId);
-  const updateBookmark = foundBookmark?.videos.filter(
+  const updateVideos = foundBookmark?.videos.filter(
     (vId) => vId.videoId.toString() !== videoId.toString(),
   );
-  foundBookmark!.count -= 1;
 
-  if (updateBookmark) foundBookmark!.videos = updateBookmark;
-  foundBookmark!.save();
+  // 북마크에서 해당 유튜브 삭제
+  if (updateVideos) {
+    foundBookmark!.count -= 1;
+    foundBookmark!.videos = updateVideos;
+    foundBookmark!.save();
+  }
 
-  await Video.findByIdAndRemove(videoId);
+  // 북마크에 남아있는 유튜브 영상이 없다면 유튜브도 삭제
+  const foundVideo = await Video.findById(videoId);
+  const result = await UserBookmarkModel.findOne(foundVideo!._id);
+  if (!result) await Video.findByIdAndRemove(videoId);
 };
 
 const User = model<IUserDocument, IUserModel>('User', userSchema);
